change default options, minor exception handling edits

+++ genomics-dataflow-core/src/main/java/com/google/allenday/genomics/core/batch/BatchProcessingPipelineOptions.java
@@ -37,7 +37,7 @@ public interface BatchProcessingPipelineOptions extends GenomicsPipelineOptions 
     void setMaxFastqSizeMB(Integer value);
 
     @Description("Size of chunk (in reads count) in which will be splited large FASTQ")
-    @Default.Integer(1000000)
+    @Default.Integer(2000000)
     Integer getMaxFastqChunkSize();
 
     void setMaxFastqChunkSize(Integer value);
@@ -49,7 +49,7 @@ public interface BatchProcessingPipelineOptions extends GenomicsPipelineOptions 
     void setFastqExt(UriProvider.FastqExt value);
 
     @Description("Num of bases that will contain region splitted BAM file")
-    @Default.Integer(10000000)
+    @Default.Integer(5000000)
     Integer getBamRegionSize();
 
     void setBamRegionSize(Integer value);
